memperbaiki passing id ke SurveyDetailActivity

+++ app/src/main/java/id/co/meda/survey/MySurveyActivity.java
@@ -11,7 +11,6 @@ import android.widget.ListView;
 import android.widget.SimpleCursorAdapter;
 
 import id.co.meda.survey.database.SurveyDatabase;
-import id.co.meda.survey.database.VoucherDatabase;
 
 public class MySurveyActivity extends ListActivity {
 
@@ -32,7 +31,7 @@ public class MySurveyActivity extends ListActivity {
     @Override
     protected void onListItemClick(ListView l, View v, int position, long id) {
         Intent intentSrveyDetail = new Intent(this, SurveyDetailActivity.class);
-        intentSrveyDetail.putExtra(SurveyDetailActivity.ID, id);
+        intentSrveyDetail.putExtra(SurveyDetailActivity.ID, (int)id);
         startActivity(intentSrveyDetail);
     }
